ref: move identifiers to separate module
